requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
hnswlib>=0.8.0
sentence-transformers>=5.2.2
numpy>=2.4.1
pandas>=3.0.0
rank-bm25>=0.2.2